requirement.txt
asgiref==3.2.10
Django==3.1.1
django-modeltranslation==0.15.2
mysqlclient==2.0.1
Pillow==7.2.0
python-gettext==4.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
